Move common structs to cluster_state.h, rename _clusterNode to clusterNode
Move clusterNode, clusterState, slotStat, slotRange struct definitions
and common node flags (PRIMARY, REPLICA, PFAIL, FAIL, MYSELF, HANDSHAKE,
NOADDR, MEET, MIGRATE_TO, NOFAILOVER) with their accessor macros from
cluster_legacy.h to cluster_state.h.

Rename struct _clusterNode to struct clusterNode, dropping the
underscore prefix.

Legacy-specific node flags (EXTENSIONS_SUPPORTED, LIGHT_HDR_*,
MULTI_MEET_SUPPORTED, NULL_NAME) and their macros remain in
cluster_legacy.h.

Update cluster_slot_stats.h to include cluster_state.h instead of
cluster_legacy.h.

Part of the cluster protocol separation refactoring to allow
alternative cluster bus implementations.
